# stats/openapi.yaml
openapi: 3.1.0
info:
  title: Carbon-Aware Forecasting Service
  description: Provides load and carbon intensity forecasts for data center locations.
    Forecasts are generated using Ridge regression models trained on historical UK
    Carbon Intensity API data and weather features from Open-Meteo. All forecasts
    cover a 7-day window at 5-minute resolution (2016 data points).
  version: 1.0.0
paths:
...
  /datacenters/{location_id}:
    patch:
      tags:
      - Datacenters
      summary: Toggle datacenter active state
      description: Activates or deactivates a datacenter. Active datacenters appear
        in /locations and receive scheduled forecasts.
      operationId: patch_datacenter_datacenters__location_id__patch
      parameters:
      - name: location_id
        in: path
        required: true
        schema:
          type: string
          title: Location Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatacenterPatchRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Datacenter'
        '404':
          description: Datacenter not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to reload datacenter after update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    Datacenter:
      properties:
        id:
          type: string
          title: Id
          description: Location identifier
          examples:
          - Data-Center-1
        region_id:
          type: integer
          title: Region Id
          description: UK Carbon Intensity API region ID
          examples:
          - 13
        name:
          type: string
          title: Name
          description: Human-readable region name
          examples:
          - London
        active:
          type: boolean
          title: Active
          description: Whether this datacenter is visible to the scheduler
        latitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Latitude
          description: Latitude for weather lookups and map display
          examples:
          - 51.51
        longitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Longitude
          description: Longitude for weather lookups and map display
          examples:
          - -0.13
      type: object
      required:
      - id
      - region_id
      - name
      - active
      title: Datacenter
    DatacenterPatchRequest:
      properties:
        active:
          type: boolean
          title: Active
          description: New active state for this datacenter
      type: object
      required:
      - active
      title: DatacenterPatchRequest
    ErrorResponse:
      properties:
        detail:
          type: string
          title: Detail
          description: Error description
      type: object
      required:
      - detail
      title: ErrorResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError